rename UT and refine UT code; test=develop
